Project/requirements.txt
asgiref==3.4.1
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.10b0
certifi==2018.10.15
chardet==3.0.4
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.2
Django==3.2.9
entrypoints==0.3
filelock==3.0.12
idna==2.7
image==1.5.33
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nest-asyncio==1.5.1
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.3.0
prompt-toolkit==3.0.22
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2018.5
pywin32==302
pyzmq==22.3.0
regex==2021.11.10
requests==2.19.1
six==1.16.0
sqlparse==0.4.2
tkcalendar==1.6.1
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.0
urllib3==1.23
virtualenv==20.7.2
wcwidth==0.2.5
